yolov5_hse/requirements.txt
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.1.0
tensorflow==2.9.0
protobuf~=3.19.0